made the datatype of counter-methods long rather than an int check some errors in the counter-methods

+++ src/SortProfiler.java
@@ -2,7 +2,7 @@
  * The {@code SortProfiler} interface defines methods for counting statements for sorting algorithms
  */
 public interface SortProfiler {
-    int getBubbleSortStatementCount(MedicalRecords[] medicalRecords);
-    int getInsertionSortStatementCount(MedicalRecords[] medicalRecords);
-    int getSelectionSortStatementCount(MedicalRecords[] medicalRecords);
+    long getBubbleSortStatementCount(MedicalRecords[] medicalRecords);
+    long getInsertionSortStatementCount(MedicalRecords[] medicalRecords);
+    long getSelectionSortStatementCount(MedicalRecords[] medicalRecords);
 }
